--- cctv_pre.csv
@@ -1,12 +1,0 @@
-,cctv_num,cctv_exp,coordinate,topLeft_x,topLeft_y,topRight_x,topRight_y,bottomRight_x,bottomRight_y,bottomLeft_x,bottomLeft_y,line_pixel,line_length,line_to_line_pixel,pred_length,real_length,오차율
-0,1,방지턱,"[[ 973. , 421.], [1061. , 422.], [1177. , 807.], [ 795. , 786.]]",973,421,1061,422,1177,807,795,786,12,15,376,470,482,0.025531915
-1,2,오토바이,"[[ 458. , 117.], [ 610. , 117.], [1000. , 650.], [ 432. , 723.]]",458,117,610,117,1000,650,432,723,15,15,414,414,413,0.002415459
-2,3,거주자우선주차,"[[ 824. ,  130.], [1002. , 130.], [1199. , 946.], [ 574. , 927.]]",824,130,1002,130,1199,946,574,927,24,15,617,385.625,394,0.02171799
-3,4,전선,"[[ 955. , 162.], [1246. , 173.], [1006. , 884.], [ 230. , 655.]]",955,162,1246,173,1006,884,230,655,17,15,775,683.8235294,691,0.010494624
-,,,,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,
