+++ web/run.py
@@ -2,4 +2,4 @@ from main import app
 
 
 if __name__ == "__main__":
-    app.run(port=8080,host='0.0.0.0', debug=True)
+    app.run(port=8080,host='0.0.0.0', debug=False)
